refactor: 删除 Core→Gateway 出站旧兼容接口并下沉出站内容模型到 common
- 将 AssistantOutput/OutputPart/OutputMedia 从 qq-maid-core 下沉到
  qq-maid-common/output_part，与入站 MessageInputPart 的分层对称
- core service重新导出这些类型，保持 qq_maid_core::service::AssistantOutput
  等对外路径稳定（common 为主，core 仅 re-export）
- 删除 CoreResponse.text / CoreResponse.markdown 旧兼容字段，
  Core→Gateway 完整回复正文只通过 output: Option<AssistantOutput> 表达
- 精简 text_content / markdown_content 访问器，只读取 output
- with_output 只设置 output，不再回填旧字段
- From<RespondResponse> for CoreResponse 直接由 RespondResponse 的
  text/markdown 中间值合成 AssistantOutput，不再向 Gateway 暴露旧字段
- 删除 AssistantOutput::from_compat_fields 旧兼容桥接（无剩余调用方）
- 迁移 core/gateway 全部测试与 fixture，删除只测试旧兼容字段的双轨测试
- Gateway render 层等出站路径继续通过统一访问器读取结构化 output，
  用户可见行为不变
