+++ qq-maid-gateway-rs/src/gateway/aggregator/tests.rs
@@ -40,8 +40,6 @@ impl CoreService for MockCore {
     async fn respond(&self, _request: CoreRequest) -> Result<CoreRespondOutput, CoreError> {
         Ok(CoreRespondOutput::Complete(Box::new(CoreResponse {
             output: Some(qq_maid_core::service::AssistantOutput::text("ok")),
-            text: Some("ok".to_owned()),
-            markdown: None,
             handled: Some(true),
             session_id: None,
             command: None,
