+++ qq-maid-gateway-rs/src/gateway/stream/tests.rs
@@ -217,8 +217,6 @@ fn c2c_message() -> C2cMessage {
 fn respond_response(text: &str) -> RespondResponse {
     RespondResponse {
         output: Some(qq_maid_core::service::AssistantOutput::markdown(text, text)),
-        text: Some(text.to_owned()),
-        markdown: Some(text.to_owned()),
         handled: Some(true),
         session_id: None,
         command: None,
